requirements.txt
opencv-python
numpy
scikit-image
scikit-learn
pandas
pyyaml
torch
torchvision
albumentations
xgboost
matplotlib
